pkg/config: extract application strict validators into compiler_validate_strict_application.go (#4405)
Pure code-motion split of the 6,997-LOC compiler_validate_strict.go
(#4405). Moves the application spec, syntax and structure strict commit-time validators
(validateXxxStrict) and their domain-local helpers verbatim into
compiler_validate_strict_application.go (package config unchanged).

No validator logic, signature, or name changed: declarations are moved
byte-for-byte and only the per-file import set is recomputed. Verified by
the sorted top-level declaration-name set staying identical across the
split, and by go build ./... plus go test ./pkg/config/... staying green
at this step. Updates the two stale doc locators that named the old file for
validateApplicationSpecsStrict.
